// src/services/models/pumpeAndPriceResponse.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * PUMPE MEME API
 * OpenAPI spec version: 1.0.0
 */
import type { UserResponse } from "./userResponse";

export interface PumpeAndPriceResponse {
  readonly address: string;
  readonly created_at: string;
  readonly createdTransaction: string;
  readonly createTime: string;
  readonly decimals: string;
  readonly description: string;
  readonly dexScreenerPairUrl: string;
  readonly factoryAddress: string;
  readonly id: number;
  readonly isWillListAfterPump: boolean;
  readonly kingOfTheHillTime: number;
  readonly logo: string;
  readonly marketCap: number;
  readonly name: string;
  readonly owner: string;
  readonly ownerInfo: UserResponse;
  readonly poolBalance: string;
  readonly poolPumped: boolean;
  readonly price24hBefore: string;
  readonly priceAfterPump: string;
  readonly priceIncreasePercentage24h: string;
  readonly priceIncreasePercentageAfterPump: string;
  readonly priceNow: string;
  readonly progressKingOfTheHill: string;
  readonly pumpTime: number;
  readonly status: string;
  readonly symbol: string;
  readonly telegram: string;
  readonly tokenId: string;
  readonly totalPumped: string;
  readonly totalRaised: string;
  readonly totalReserve: string;
  readonly totalSupply: string;
  readonly uniswapPair: string;
  readonly uniswapV2Router: string;
  readonly updated_at: string;
  readonly volume1h: number;
  readonly volume24hs: number;
  readonly website: string;
  readonly x: string;
}