// src/services/models/pumpeControllerFindPaginationSortBy.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * PUMPE MEME API
 * OpenAPI spec version: 1.0.0
 */

export type PumpeControllerFindPaginationSortBy =
  (typeof PumpeControllerFindPaginationSortBy)[keyof typeof PumpeControllerFindPaginationSortBy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PumpeControllerFindPaginationSortBy = {
  createTime: "createTime",
  volume1h: "volume1h",
  volume24hs: "volume24hs",
  marketCap: "marketCap",
  totalVoting: "totalVoting",
} as const;