// src/services/models/pumpeControllerFindPaginationStatus.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * PUMPE MEME API
 * OpenAPI spec version: 1.0.0
 */

export type PumpeControllerFindPaginationStatus =
  (typeof PumpeControllerFindPaginationStatus)[keyof typeof PumpeControllerFindPaginationStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PumpeControllerFindPaginationStatus = {
  active: "active",
  hidden: "hidden",
  voting: "voting",
  upcoming: "upcoming",
  completed: "completed",
} as const;